+++ frontend/src/components/project-wizard/house-blocks/ownership-information/OwnershipInformationGroup.tsx
@@ -1,4 +1,4 @@
-import { Grid, IconButton, Stack, Typography } from "@mui/material";
+import { Grid, IconButton, Typography } from "@mui/material";
 import { WizardCard } from "../../WizardCard";
 import { t } from "i18next";
 import { HouseBlock, OwnershipSingleValue } from "../../../../types/houseBlockTypes";
@@ -21,8 +21,8 @@ export const OwnershipInformationGroup = ({ houseBlock, setHouseBlock, readOnly 
                 {
                     type: undefined,
                     amount: 0,
-                    value: { value: 0, min: null, max: null },
-                    rentalValue: { value: 0, min: null, max: null },
+                    value: { value: null, min: null, max: null },
+                    rentalValue: { value: null, min: null, max: null },
                 },
             ],
         });
@@ -52,22 +52,23 @@ export const OwnershipInformationGroup = ({ houseBlock, setHouseBlock, readOnly 
             <Typography fontWeight={600} mb={2}>
                 {t(`${translationPath}.ownershipAndValue.title`)}
             </Typography>
-            <Stack direction="row" alignItems="center" spacing={2} my={1}>
-                <Typography fontWeight={600} flex={4.2}>
-                    {t(`${translationPath}.type`)}
-                    <TooltipInfo text={t(`tooltipInfo.soort.title`)} />
-                </Typography>
-                <Typography fontWeight={600} flex={2}>
-                    {t(`${translationPath}.amount`)}
-                </Typography>
-                <Typography fontWeight={600} flex={2}>
-                    {t(`${translationPath}.value`)}
-                </Typography>
-                <Typography fontWeight={600} flex={2}>
-                    {t(`${translationPath}.rent`)}
-                </Typography>
-                <Typography fontWeight={600} flex={2}></Typography>
-            </Stack>
+            <Grid container spacing={2}>
+                <Grid item xs={4}>
+                    <Typography fontWeight={600}>
+                        {t(`${translationPath}.type`)}
+                        <TooltipInfo text={t(`tooltipInfo.soort.title`)} />
+                    </Typography>
+                </Grid>
+                <Grid item xs={1.25}>
+                    <Typography fontWeight={600}>{t(`${translationPath}.amount`)}</Typography>
+                </Grid>
+                <Grid item xs={2.8}>
+                    <Typography fontWeight={600}>{t(`${translationPath}.value`)}</Typography>
+                </Grid>
+                <Grid item xs={2.8}>
+                    <Typography fontWeight={600}>{t(`${translationPath}.rent`)}</Typography>
+                </Grid>
+            </Grid>
             <Grid container>
                 {houseBlock.ownershipValue.map((ownership, index) => (
                     <OwnershipRowInputs